+++ src/generator.rs
@@ -183,30 +183,16 @@ impl<R: rand::RngCore> Scru128Generator<R> {
     ///
     /// Note that the generator object should be protected from concurrent accesses during the
     /// sequential calls to a generation method and this method to avoid race conditions.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # #[cfg(feature = "std")]
-    /// # {
-    /// use scru128::generator::{Scru128Generator, Status};
-    ///
-    /// let mut g = Scru128Generator::new();
-    /// let x = g.generate();
-    /// let y = g.generate();
-    /// if g.last_status() == Status::ClockRollback {
-    ///     panic!("clock moved backward");
-    /// } else {
-    ///     assert!(x < y);
-    /// }
-    /// # }
-    /// ```
+    #[deprecated(
+        since = "2.6.0",
+        note = "use `generate_monotonic()` to guarantee monotonicity"
+    )]
     pub const fn last_status(&self) -> Status {
         self.last_status
     }
 }
 
-/// Status code returned by [`Scru128Generator::last_status()`] method.
+/// _Deprecated_. Status code returned by [`Scru128Generator::last_status()`] method.
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
 pub enum Status {
     /// Indicates that the generator has yet to generate an ID.
@@ -323,18 +309,18 @@ mod tests_generate_core {
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_core(ts);
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         for i in 0..100_000u64 {
             let curr = g.generate_core(ts - i.min(9_998));
             assert!(
-                g.last_status() == Status::CounterLoInc
-                    || g.last_status() == Status::CounterHiInc
-                    || g.last_status() == Status::TimestampInc
+                g.last_status == Status::CounterLoInc
+                    || g.last_status == Status::CounterHiInc
+                    || g.last_status == Status::TimestampInc
             );
             assert!(prev < curr);
             prev = curr;
@@ -347,23 +333,23 @@ mod tests_generate_core {
     fn breaks_increasing_order_of_ids_if_timestamp_moves_backward_a_lot() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_core(ts);
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         let mut curr = g.generate_core(ts - 10_000);
-        assert_eq!(g.last_status(), Status::ClockRollback);
+        assert_eq!(g.last_status, Status::ClockRollback);
         assert!(prev > curr);
         assert_eq!(curr.timestamp(), ts - 10_000);
 
         prev = curr;
         curr = g.generate_core(ts - 10_001);
         assert!(
-            g.last_status() == Status::CounterLoInc
-                || g.last_status() == Status::CounterHiInc
-                || g.last_status() == Status::TimestampInc
+            g.last_status == Status::CounterLoInc
+                || g.last_status == Status::CounterHiInc
+                || g.last_status == Status::TimestampInc
         );
         assert!(prev < curr);
     }
@@ -379,18 +365,18 @@ mod tests_generate_core_monotonic {
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_core_monotonic(ts).unwrap();
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         for i in 0..100_000u64 {
             let curr = g.generate_core_monotonic(ts - i.min(9_998)).unwrap();
             assert!(
-                g.last_status() == Status::CounterLoInc
-                    || g.last_status() == Status::CounterHiInc
-                    || g.last_status() == Status::TimestampInc
+                g.last_status == Status::CounterLoInc
+                    || g.last_status == Status::CounterHiInc
+                    || g.last_status == Status::TimestampInc
             );
             assert!(prev < curr);
             prev = curr;
@@ -403,19 +389,19 @@ mod tests_generate_core_monotonic {
     fn returns_none_if_timestamp_moves_backward_a_lot() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let prev = g.generate_core_monotonic(ts).unwrap();
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         let mut curr = g.generate_core_monotonic(ts - 10_000);
         assert!(curr.is_none());
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
 
         curr = g.generate_core_monotonic(ts - 10_001);
         assert!(curr.is_none());
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
     }
 }